Move assets outside the APK, add file helper functions.
Added a File module which provides helper functions to read/write from files.
It uses platform-specific paths to the assets and the temporary directories.

Updated CMake to use adb push --sync to push files that are newer
on the host than the device (requires version > 1.0.40).

Removed all references to Android's Asset Manager, since now all assets are
synced to the target device storage outside the APK.

Moved spdlog initialization to platform constructor so that it can be used
at platform initialization.