Framework changes for command buffer features
These changes allow the framework to coordinate RenderContext, RenderTarget and CommandPool
on how to allocate and manage command buffers.
For example, they could be reset together with the pool, or individually, or both reset and freed.

They also allow samples to record to secondary command buffers,
which required changes to CommandRecord and CommandReplay so that the secondary command buffers
correctly inherit the necessary information from the primary command buffer.

Added a state attribute to command buffers to track their lifecycle.

Refactored SceneSubpass so that samples can call draw_submesh
if they choose a different order for drawing the scene meshes.

Refactored VulkanSample so that samples can more easily override update
and draw_swapchain_renderpass.